moved classes related to encoding to package read.encoding
