+++ connector/src/main/scala/io/github/dejarol/azure/search/spark/connector/core/schema/SchemaUtils.scala
@@ -2,7 +2,6 @@ package io.github.dejarol.azure.search.spark.connector.core.schema
 
 import com.azure.search.documents.indexes.models.{SearchField, SearchFieldDataType}
 import io.github.dejarol.azure.search.spark.connector.core.{DataTypeException, JavaScalaConverters}
-import io.github.dejarol.azure.search.spark.connector.core.schema.conversion.GeoPointType
 import org.apache.spark.sql.types._
 
 import java.util.{List => JList}
@@ -42,7 +41,7 @@ object SchemaUtils {
     } else if (searchFieldDataType.isComplex) {
       inferSparkComplexType(searchField.getFields)
     } else if (searchFieldDataType.isGeoPoint) {
-      GeoPointType.SCHEMA
+      GeoPointType.SPARK_SCHEMA
     } else {
       throw DataTypeException.forUnsupportedSearchType(searchField.getType)
     }
@@ -194,7 +193,7 @@ object SchemaUtils {
 
     evaluateSparkTypesCompatibility(
       dataType,
-      GeoPointType.SCHEMA
+      GeoPointType.SPARK_SCHEMA
     )
   }
