+++ connector/src/main/scala/io/github/dejarol/azure/search/spark/connector/read/SearchPartitionReaderFactory.scala
@@ -1,7 +1,8 @@
 package io.github.dejarol.azure.search.spark.connector.read
 
-import io.github.dejarol.azure.search.spark.connector.core.schema.CodecCreationException
+import io.github.dejarol.azure.search.spark.connector.core.codec.CodecCreationException
 import io.github.dejarol.azure.search.spark.connector.read.config.ReadConfig
+import io.github.dejarol.azure.search.spark.connector.read.encoding.SearchDocumentEncoderImpl
 import io.github.dejarol.azure.search.spark.connector.read.partitioning.SearchPartition
 import org.apache.spark.sql.catalyst.InternalRow
 import org.apache.spark.sql.connector.read.{InputPartition, PartitionReader, PartitionReaderFactory}
