+++ connector/src/test/scala/io/github/dejarol/azure/search/spark/connector/core/schema/conversion/input/ComplexEncoderSpec.scala
@@ -1,7 +1,8 @@
 package io.github.dejarol.azure.search.spark.connector.core.schema.conversion.input
 
 import io.github.dejarol.azure.search.spark.connector.BasicSpec
-import io.github.dejarol.azure.search.spark.connector.core.schema.conversion.SearchIndexColumnImpl
+import io.github.dejarol.azure.search.spark.connector.core.codec.SearchIndexColumnImpl
+import io.github.dejarol.azure.search.spark.connector.read.encoding.{AtomicEncoders, ComplexEncoder}
 import org.apache.spark.sql.catalyst.InternalRow
 import org.apache.spark.sql.types.DataTypes
