+++ connector/src/test/scala/io/github/dejarol/azure/search/spark/connector/core/schema/conversion/output/StructTypeDecoderSpec.scala
@@ -1,7 +1,7 @@
 package io.github.dejarol.azure.search.spark.connector.core.schema.conversion.output
 
 import io.github.dejarol.azure.search.spark.connector.BasicSpec
-import io.github.dejarol.azure.search.spark.connector.core.schema.conversion.SearchIndexColumnImpl
+import io.github.dejarol.azure.search.spark.connector.core.codec.SearchIndexColumnImpl
 import org.apache.spark.sql.catalyst.InternalRow
 import org.apache.spark.sql.types.DataTypes
