+++ connector/src/test/scala/io/github/dejarol/azure/search/spark/connector/read/EncoderFactorySpec.scala
@@ -2,7 +2,9 @@ package io.github.dejarol.azure.search.spark.connector.read
 
 import com.azure.search.documents.indexes.models.SearchFieldDataType
 import io.github.dejarol.azure.search.spark.connector.core.Constants
-import io.github.dejarol.azure.search.spark.connector.core.schema.{CodecErrors, CodecFactorySpec}
+import io.github.dejarol.azure.search.spark.connector.core.codec.CodecErrors
+import io.github.dejarol.azure.search.spark.connector.core.schema.CodecFactorySpec
+import io.github.dejarol.azure.search.spark.connector.read.encoding.EncoderFactory
 import org.apache.spark.sql.types.{DataType, DataTypes}
 import org.apache.spark.unsafe.types.UTF8String
